refactor: reorganize test helpers and update imports for consistency
