+++ programs/solana/tests/src/lib.rs
@@ -1,8 +1,5 @@
 pub mod helpers;
 #[cfg(test)]
-mod test_helpers;
-
-#[cfg(test)]
 mod test_initialize;
 
 #[cfg(test)]
