+++ programs/solana/tests/src/test_verify_non_membership.rs
@@ -1,4 +1,4 @@
-// use crate::common::with_initialized_client;
+// use crate::helpers::with_initialized_client;
 // use crate::test_helpers::create_test_merkle_proof_bytes;
 // use anchor_client::solana_sdk::{signer::Signer, pubkey::Pubkey};
 // use ics07_tendermint::MembershipMsg;
